tests/wpt/web-platform-tests/css/css-tables/parsing/caption-side-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing caption-side with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-caption-side">
<meta name="assert" content="caption-side supports only the grammar 'top | bottom'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("caption-side", "auto");
test_invalid_value("caption-side", "left");
test_invalid_value("caption-side", "right");
test_invalid_value("caption-side", "top bottom");
</script>
</body>
</html>
